RIA-1935 add addendum evidence HO
