+++ src/functionalTest/resources/scenarios/RIA-1935-upload-addendum-evidence-ho-submit.json
@@ -1,0 +1,112 @@
+{
+  "description": "RIA-1935 Upload addendum evidence Home Office",
+  "request": {
+    "uri": "/asylum/ccdAboutToSubmit",
+    "credentials": "HomeOfficeGeneric",
+    "input": {
+      "id": 1234,
+      "eventId": "uploadAddendumEvidenceHomeOffice",
+      "state": "preHearing",
+      "caseData": {
+        "template": "minimal-appeal-submitted.json",
+        "replacements": {
+          "addendumEvidenceDocuments": [
+            {
+              "id": "1",
+              "value": {
+                "document": {
+                  "document_url": "http://document-store/AAA",
+                  "document_binary_url": "http://document-store/AAA/binary",
+                  "document_filename": "existing-evidence.pdf"
+                },
+                "description": "Existing evidence",
+                "dateUploaded": "2018-12-25",
+                "tag": "addendumEvidence",
+                "suppliedBy": "The appellant"
+              }
+            }
+          ],
+          "addendumEvidence": [
+            {
+              "id": "1",
+              "value": {
+                "document": {
+                  "document_url": "http://document-store/BBB",
+                  "document_binary_url": "http://document-store/BBB/binary",
+                  "document_filename": "some-new-evidence.pdf"
+                },
+                "description": "Some new evidence"
+              }
+            },
+            {
+              "id": "2",
+              "value": {
+                "document": {
+                  "document_url": "http://document-store/CCC",
+                  "document_binary_url": "http://document-store/CCC/binary",
+                  "document_filename": "some-more-new-evidence.pdf"
+                },
+                "description": "Some more new evidence"
+              }
+            }
+          ],
+          "uploadAddendumEvidenceHomeOfficeActionAvailable": "Yes"
+        }
+      }
+    }
+  },
+  "expectation": {
+    "status": 200,
+    "errors": [],
+    "caseData": {
+      "template": "minimal-appeal-submitted.json",
+      "replacements": {
+        "addendumEvidenceDocuments": [
+          {
+            "id": "3",
+            "value": {
+              "document": {
+                "document_url": "http://document-store/BBB",
+                "document_binary_url": "http://document-store/BBB/binary",
+                "document_filename": "some-new-evidence.pdf"
+              },
+              "description": "Some new evidence",
+              "dateUploaded": "{$TODAY}",
+              "tag": "addendumEvidence",
+              "suppliedBy": "The respondent"
+            }
+          },
+          {
+            "id": "2",
+            "value": {
+              "document": {
+                "document_url": "http://document-store/CCC",
+                "document_binary_url": "http://document-store/CCC/binary",
+                "document_filename": "some-more-new-evidence.pdf"
+              },
+              "description": "Some more new evidence",
+              "dateUploaded": "{$TODAY}",
+              "tag": "addendumEvidence",
+              "suppliedBy": "The respondent"
+            }
+          },
+          {
+            "id": "1",
+            "value": {
+              "document": {
+                "document_url": "http://document-store/AAA",
+                "document_binary_url": "http://document-store/AAA/binary",
+                "document_filename": "existing-evidence.pdf"
+              },
+              "description": "Existing evidence",
+              "dateUploaded": "2018-12-25",
+              "tag": "addendumEvidence",
+              "suppliedBy": "The appellant"
+            }
+          }
+        ],
+        "uploadAddendumEvidenceHomeOfficeActionAvailable": "Yes"
+      }
+    }
+  }
+}
