+++ src/test/java/uk/gov/hmcts/reform/iacaseapi/domain/handlers/postsubmit/UploadAddendumEvidenceConfirmationTest.java
@@ -65,7 +65,9 @@ public class UploadAddendumEvidenceConfirmationTest {
 
             boolean canHandle = uploadAddendumEvidenceConfirmation.canHandle(callback);
 
-            if (event == Event.UPLOAD_ADDENDUM_EVIDENCE || event == Event.UPLOAD_ADDENDUM_EVIDENCE_LEGAL_REP) {
+            if (event == Event.UPLOAD_ADDENDUM_EVIDENCE
+                || event == Event.UPLOAD_ADDENDUM_EVIDENCE_LEGAL_REP
+                || event == Event.UPLOAD_ADDENDUM_EVIDENCE_HOME_OFFICE) {
 
                 assertTrue(canHandle);
             } else {
